Rename util package name and file
