+++ internal/fsutil/helpers.go
@@ -1,4 +1,4 @@
-package util
+package fsutil
 
 import (
 	"io/fs"
